Addressing Issue #25, SLES 15 SP6 compile failure
Changed the linux-mdep.c to increase level check from greater than 6.0.0. to 6.4.0 for the i_time, a_time and c_time values for time of day usage. 
Also, updated the version.c to list '05 Dec 24' as the date of the change and the version is 'CS7111A'.  You can verify the version level by using:
'ps -elf | grep LiS' and verify CS7111A is listed as the process name for the LiS dedicated queues on the CPUs.