Fix RHEL 9.5 build error on ppc64le and x86_64
Error in build on RHEL 9.5 for ppc64le and x86_64 release for missing invalidate_device(). The new call in invalidate_bdev(). 
Changed the linux-mdep.c to check for RHEL 9.5 level for ppc64le and x86_64, while RHEL 9.4 on s390x is checked at 5.14.0 level. 
Also, updated the version.c to list '13 Dec 24' as the date of the change and the version is 'CS7111B'.  You can verify the version level by using:
'ps -elf | grep LiS' and verify CS7111B is listed as the process name for the LiS dedicated queues on the CPUs.