RHEL 9.5 ppc64le and x86_64 build fix.
This package contains fixes to ppc64le and x86_64 platform builds where a Linux 6 kernel change was implemented on 5.14.0. For RHEL 9.4, s390x, this same change was made. Due to inconsistent code release by RHEL, this fix was needed to cover the proper platforms release levels. 
The version.c to list '13 Dec 24' as the date of the change and the version is 'CS7111B'.  You can verify the version level by using:
'ps -elf | grep LiS' and verify CS7111B is listed as the process name for the LiS dedicated queues on the CPUs.
Replaces the previous fix release CSLiS-7.1.1.1.241205.tgz.